+++ CHANGELOG.md
@@ -5,6 +5,17 @@ Document tracking all changes made to the yaqcThor package.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2023-10-10
+
+### Changed
+
+- Modified `helpers.waitfor` to work with indefinite waiting time
+- Fixed runtime errors associated with `experiments.PLE_spectrum`
+
+### Removed
+
+- Removed objective scales from `constants` module in deference to `makeitwright.andor` method `set_scale`
+
 ## [0.1.0] - 2023-01-18
 
 ### Added